// front-end/src/constants/types.ts
enum ACTION_TYPES {
    OPERATION_APPEND = "OPERATION_APPEND",
    OPERATION_UPDATE = "OPERATION_UPDATE",
    OPERATION_DELETE = "OPERATION_DELETE",
    OPERATION_MOVE_UP = "OPERATION_MOVE_UP",
    OPERATION_MOVE_DOWN = "OPERATION_MOVE_DOWN",
    OPERATION_CLEAR = "OPERATION_CLEAR",
    OPERATION_SUBSTITUTE_VARIABLE = "OPERATION_SUBSTITUTE_VARIABLE",
    RESULT_APPEND = "RESULT_APPEND",
    RESULT_DELETE = "RESULT_DELETE",
    RESULT_CLEAR = "RESULT_CLEAR",
    RESULT_SHOW_LOADER = "RESULT_SHOW_LOADER",
    RESULT_SHOW_ERROR = "RESULT_SHOW_ERROR",
    RESULT_HIDE_LOADER = "RESULT_HIDE_LOADER",
    RESULT_HIDE_ERROR = "RESULT_HIDE_ERROR",
};

enum INPUT_TYPES {
    TEXT = 'TEXT',
    SELECT = 'SELECT',
    BUTTON = 'BUTTON',
    VARIABLE_TEXT = 'VARIABLE_TEXT',
    VARIABLE_BOX = 'VARIABLE_BOX',
    CONDITION = 'CONDITION',
    OPERATIONS = "OPERATIONS"
};

enum MODES_TYPES {
    SET = "SET",
    APPEND = "APPEND"
};

enum CATEGORY_TYPES {
    ACTION = 'ACTION',
    CONDITION = 'CONDITION',
    LOOP = 'LOOP'
};

enum OPERATION_TYPES {
    OPEN = 'OPEN',
    EXTRACT = 'EXTRACT',
    CLICK = 'CLICK',
    TYPE = 'TYPE',
    SET = 'SET',
    IF = 'IF',
    IF_ELSE = 'IF_ELSE',
    WHILE = 'WHILE',
    FOR = 'FOR',
};

enum VALIDATION_TYPES {
    REQUIRED = "REQUIRED",
    URL = "URL",
    EMAIL = "EMAIL"
};

export { ACTION_TYPES, INPUT_TYPES, MODES_TYPES, CATEGORY_TYPES, OPERATION_TYPES, VALIDATION_TYPES };